web-app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
groq>=0.5.0
pydantic>=2.6.0
openai>=1.0.0

# Celery & Redis
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# MongoDB
pymongo>=4.6.0

# MinIO (S3)
minio>=7.2.0

# Google YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.0
